+++ source/FLDN_API/Infrastructure/Data/SeedData/AppData.cs
@@ -1,3 +1,6 @@
+using Domain;
+using Microsoft.EntityFrameworkCore;
+
 namespace Infrastructure;
 
 public class AppData
@@ -5,8 +8,9 @@ public class AppData
     public static async Task SeedAsync(ApplicationDbContext context)
     {
         if (!await context.Roles.AnyAsync())
+        {
             await context.Roles.AddRangeAsync(RoleData.GetRoles());
-
-        await context.SaveChangesAsync();
+            await context.SaveChangesAsync();
+        }
     }
 }